requirements-dev.txt
aiofiles==0.6.0; python_version >= "3.5"
aiohttp-swagger==1.0.15; python_version >= "3.5"
aiohttp==3.6.3; python_full_version >= "3.5.3" and python_version >= "3.5"
appdirs==1.4.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
astroid==2.5.2; python_version >= "3.6" and python_version < "4.0"
async-lru==1.0.2; python_version >= "3.5"
async-timeout==3.0.1; python_full_version >= "3.5.3" and python_version >= "3.5"
attrs==20.3.0; python_full_version >= "3.5.3" and python_version >= "3.5"
autopep8==1.5.6
black==20.8b1; python_version >= "3.6"
certifi==2020.12.5; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
cffi==1.14.5; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
chardet==3.0.4; python_full_version >= "3.5.3" and python_version >= "3.6"
click==7.1.2; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
colorama==0.4.4; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" and sys_platform == "win32" and platform_system == "Windows" or python_version >= "3.6" and python_version < "4.0" and sys_platform == "win32" and python_full_version >= "3.5.0" and platform_system == "Windows"
cryptography==3.3.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
deprecated==1.2.12; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.6"
dill==0.3.3; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.1.0" and python_version >= "3.5"
distlib==0.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
distro==1.5.0; python_version >= "3.5"
filelock==3.0.12; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
flake8==3.9.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
future==0.18.2; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.3.0" and python_version >= "3.5"
idna==2.10; python_full_version >= "3.5.3" and python_version >= "3.6"
importlib-metadata==3.10.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.5.0" and python_version < "3.8" and python_version >= "3.6"
isort==5.8.0; python_version >= "3.6" and python_version < "4.0"
jinja2==2.11.2; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
joblib==1.0.1; python_version >= "3.6"
junit-xml==1.9; python_version >= "3.5"
lazy-object-proxy==1.6.0; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.6.0"
livereload==2.6.3; python_version >= "3.5"
lunr==0.5.8; python_version >= "3.5"
markdown==3.3.4; python_version >= "3.6"
markupsafe==1.1.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.5"
mccabe==0.6.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.5.0"
mkdocs==1.1.2; python_version >= "3.5"
multidict==4.7.6; python_full_version >= "3.5.3" and python_version >= "3.5"
mypy-extensions==0.4.3; python_version >= "3.6"
nltk==3.5; python_version >= "3.5"
packaging==20.9; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pathspec==0.8.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
pluggy==0.13.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
psutil==5.8.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.4.0" and python_version >= "3.5"
py==1.10.0; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pyats.aereport==21.2; python_version >= "3.5"
pyats.aetest==21.2; python_version >= "3.5"
pyats.async==21.2; python_version >= "3.5"
pyats.connections==21.2.1; python_version >= "3.5"
pyats.datastructures==21.2; python_version >= "3.5"
pyats.easypy==21.2; python_version >= "3.5"
pyats.kleenex==21.2; python_version >= "3.5"
pyats.log==21.2; python_version >= "3.5"
pyats.reporter==21.2; python_version >= "3.5"
pyats.results==21.2; python_version >= "3.5"
pyats.tcl==21.2; python_version >= "3.5"
pyats.topology==21.2; python_version >= "3.5"
pyats.utils==21.2; python_version >= "3.5"
pyats==21.2; python_version >= "3.5"
pycodestyle==2.7.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.4.0")
pycparser==2.20; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
pyflakes==2.3.1; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
pygithub==1.54.1; python_version >= "3.6"
pyjwt==1.7.1; python_version >= "3.6"
pylint==2.7.4; python_version >= "3.6" and python_version < "4.0"
pyparsing==2.4.7; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
python-engineio==3.13.2; python_version >= "3.5"
python-socketio==4.6.0; python_version >= "3.5"
pyyaml==5.4.1; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
regex==2021.3.17; python_version >= "3.6"
requests==2.25.1; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version >= "3.6"
robotframework==4.0
six==1.15.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_full_version >= "3.6.0" and python_version >= "3.5"
toml==0.10.2; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_version >= "3.6" and python_version < "4.0" and python_full_version >= "3.5.0"
tornado==6.1; python_version >= "3.5"
tox==3.23.0; (python_version >= "2.7" and python_full_version < "3.0.0") or (python_full_version >= "3.5.0")
tqdm==4.59.0; python_version >= "3.5" and python_full_version < "3.0.0" or python_version >= "3.5" and python_full_version >= "3.4.0"
typed-ast==1.4.2; implementation_name == "cpython" and python_version < "3.8" and python_version >= "3.6"
typing-extensions==3.7.4.3; python_full_version >= "3.5.3" and python_version < "3.8" and python_version >= "3.6"
unicon.plugins==21.2; python_version >= "3.5"
unicon==21.2; python_version >= "3.5"
urllib3==1.26.4; python_version >= "3.6" and python_full_version < "3.0.0" or python_full_version >= "3.5.0" and python_version < "4" and python_version >= "3.6"
virtualenv==20.4.3; python_version >= "2.7" and python_full_version < "3.0.0" or python_full_version >= "3.5.0"
wrapt==1.12.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "4.0" or python_full_version >= "3.4.0" and python_version >= "3.6" and python_version < "4.0"
yamllint==1.26.0; python_version >= "3.5"
yarl==1.5.1; python_full_version >= "3.5.3" and python_version >= "3.5"
zipp==3.4.1; python_version >= "3.6" and python_full_version < "3.0.0" and python_version < "3.8" or python_full_version >= "3.5.0" and python_version < "3.8" and python_version >= "3.6"
